+++ src/model/directive.rs
@@ -74,14 +74,17 @@ pub struct __Directive<'a> {
 /// In some cases, you need to provide options to alter GraphQL's execution behavior in ways field arguments will not suffice, such as conditionally including or skipping a field. Directives provide this by describing additional information to the executor.
 #[Object(internal, name = "__Directive")]
 impl<'a> __Directive<'a> {
+    #[inline]
     async fn name(&self) -> &str {
         self.directive.name
     }
 
+    #[inline]
     async fn description(&self) -> Option<&str> {
         self.directive.description
     }
 
+    #[inline]
     async fn locations(&self) -> &Vec<__DirectiveLocation> {
         &self.directive.locations
     }