+++ src/model/field.rs
@@ -9,10 +9,12 @@ pub struct __Field<'a> {
 /// Object and Interface types are described by a list of Fields, each of which has a name, potentially a list of arguments, and a return type.
 #[Object(internal, name = "__Field")]
 impl<'a> __Field<'a> {
+    #[inline]
     async fn name(&self) -> &str {
         &self.field.name
     }
 
+    #[inline]
     async fn description(&self) -> Option<&str> {
         self.field.description
     }
@@ -37,11 +39,13 @@ impl<'a> __Field<'a> {
         __Type::new(self.registry, &self.field.ty)
     }
 
+    #[inline]
     async fn is_deprecated(&self) -> bool {
-        self.field.deprecation.is_some()
+        self.field.deprecation.is_deprecated()
     }
 
+    #[inline]
     async fn deprecation_reason(&self) -> Option<&str> {
-        self.field.deprecation
+        self.field.deprecation.reason()
     }
 }